\name{viz14-package}
\alias{viz14-package}
\alias{viz14}
\docType{package}
\title{
visualization package for CSAMA Brixen 2014
}
\description{
visualization package for CSAMA Brixen 2014
}
\details{
\tabular{ll}{
Package: \tab viz14\cr
Version: \tab 0.0.0\cr
Suggests: \tab \cr
Depends: \tab ggplot2, ggbio, yeastCC, RNAseqData.HNRNPC.bam.chr14\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
Built: \tab R 3.1.0; ; 2014-06-02 18:30:20 UTC; unix\cr
}
moderately documented utilities and vignettes related to
visualization
}
\author{
VJ Carey <stvjc@channing.harvard.edu>

Maintainer: VJ Carey <stvjc@channing.harvard.edu>
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%~~ documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
